Adds the 'escape-html' transformer.
It escapes the &, <, >, ", and ' characters into HTML entities so that all HTML in the passed string will be interpreted as plain text by the browser. This is useful for showing things like example HTML snippets of code or arbitrary user input (that you don't want to be interpreted as HTML).